refactor canViewAdmin
