+++ tdrs-frontend/src/components/Header/Header.jsx
@@ -3,6 +3,7 @@ import { useSelector } from 'react-redux'
 import closeIcon from 'uswds/dist/img/close.svg'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faSignOutAlt, faUserCircle } from '@fortawesome/free-solid-svg-icons'
+import canViewAdmin from '../../utils/canViewAdmin'
 
 import NavItem from '../NavItem/NavItem'
 
@@ -21,20 +22,12 @@ function Header() {
   const user = useSelector((state) => state.auth.user)
   const authenticated = useSelector((state) => state.auth.authenticated)
   const userAccessRequestPending = Boolean(user?.['access_request'])
-  const userAccessRequestApproved =
-    Boolean(user?.['access_request']) && user.roles.length > 0
-
-  const isMemberOfOne = (...groupNames) =>
-    user?.roles?.some((role) => groupNames.includes(role.name))
 
   const hasPermission = (permissionName) =>
     user?.roles?.[0]?.permissions?.some(
       (perm) => perm.codename === permissionName
     )
 
-  const canViewAdmin =
-    userAccessRequestApproved &&
-    isMemberOfOne('Developer', 'OFA System Admin', 'ACF OCIO')
   const canViewDataFiles = hasPermission('view_datafile')
 
   const menuRef = useRef()
@@ -134,7 +127,7 @@ function Header() {
                       href="/profile"
                     />
                   )}
-                  {canViewAdmin && (
+                  {canViewAdmin(user) && (
                     <NavItem
                       pathname={pathname}
                       tabTitle="Admin"
