+++ tdrs-frontend/src/components/SiteMap/SiteMap.jsx
@@ -1,17 +1,8 @@
 import React from 'react'
-
+import canViewAdmin from '../../utils/canViewAdmin'
 // I copied these from Header.jsx and slightly refactored them to accept a user as their first arguement
 // I am wanting to shift these, and similar reused pieces of code at the top of component functions
 // up the tree to be passed in from the router.
-const isMemberOfOne = (user, ...groupNames) =>
-  !!user?.roles?.some((role) => groupNames.includes(role.name))
-
-const userAccessRequestApproved = (user) =>
-  user?.['access_request'] && user?.roles?.length > 0
-
-const canViewAdmin = (user) =>
-  userAccessRequestApproved(user) &&
-  isMemberOfOne(user, 'Developer', 'OFA System Admin', 'ACF OCIO')
 
 const SiteMap = ({ user }) => (
   <div className="margin-top-5">
